Add M5Sensor file
This is the file needed to get the distance from the new sensors we got. it should work and shouldn't need to be changed. 

It communicates with the sensor directly in order to get the distance, because the sensor isnt directly supported by rev/ftc/the hub. 

its main methods are: 
init() - get the sensor and set some startup variables
update() - tells the sensor to get a new distance
getDistanceCm() - returns the last cm distance